add receipts_root verify & move the verify_block_integrity
