+++ bin/stateless-validator/Cargo.toml
@@ -9,12 +9,10 @@ path = "src/main.rs"
 
 [dependencies]
 # alloy
-alloy-consensus.workspace = true
 alloy-genesis.workspace = true
 alloy-primitives.workspace = true
 alloy-provider.workspace = true
 alloy-rpc-types-eth.workspace = true
-alloy-trie.workspace = true
 
 # op
 op-alloy-network.workspace = true
