+++ bin/stateless-validator/src/rpc.rs
@@ -1,16 +1,14 @@
 //! RPC client for fetching missing data during stateless validation.
-use alloy_consensus::transaction::SignerRecoverable;
 use alloy_primitives::{Address, B256, Bytes};
 use alloy_provider::{Provider, ProviderBuilder, RootProvider};
-use alloy_rpc_types_eth::{Block, BlockId, BlockNumberOrTag, BlockTransactions};
-use alloy_trie::root::ordered_trie_root_with_encoder;
+use alloy_rpc_types_eth::{Block, BlockId, BlockNumberOrTag};
 use eyre::{Context, Result, ensure, eyre};
 use futures::future::try_join_all;
-use op_alloy_network::{Optimism, TransactionResponse, eip2718::Encodable2718};
+use op_alloy_network::Optimism;
 use op_alloy_rpc_types::Transaction;
 use salt::SaltWitness;
 use serde::{Deserialize, Serialize};
-use validator_core::withdrawals::MptWitness;
+use validator_core::{executor::verify_block_integrity, withdrawals::MptWitness};
 
 /// Response from mega_setValidatedBlocks RPC call
 #[derive(Debug, Clone, Serialize, Deserialize)]
@@ -126,46 +124,7 @@ impl RpcClient {
             _ => {} // Skip for latest, earliest, pending, etc.
         }
 
-        // Verify block hash matches the computed hash from header
-        ensure!(
-            block.header.hash_slow() == block.header.hash,
-            "Block hash mismatch: expected {:?}, computed {:?}",
-            block.header.hash,
-            block.header.hash_slow()
-        );
-
-        // Verify transaction hashes and transactions root
-        if let BlockTransactions::Full(ref transactions) = block.transactions {
-            for tx in transactions {
-                let tx_envelope = tx.inner.clone().into_inner();
-                ensure!(
-                    tx_envelope.trie_hash() == tx.tx_hash(),
-                    "Transaction hash mismatch: expected {:?}, computed {:?}",
-                    tx.tx_hash(),
-                    tx_envelope.trie_hash()
-                );
-
-                let recovered = tx_envelope
-                    .recover_signer()
-                    .map_err(|err| eyre!("Failed to recover signer: {}", err))?;
-
-                ensure!(
-                    recovered == tx.from(),
-                    "Transaction signer mismatch: expected {:?}, got {:?}",
-                    tx.from(),
-                    recovered
-                );
-            }
-            let computed_tx_root = ordered_trie_root_with_encoder(transactions, |tx, buf| {
-                tx.inner.clone().into_inner().encode_2718(buf)
-            });
-            ensure!(
-                computed_tx_root == block.header.transactions_root,
-                "Transactions root mismatch: expected {:?}, computed {:?}",
-                block.header.transactions_root,
-                computed_tx_root
-            );
-        }
+        verify_block_integrity(&block)?;
 
         Ok(block)
     }
