+++ validator-core/Cargo.toml
@@ -10,15 +10,16 @@ alloy-eips.workspace = true
 alloy-evm.workspace = true
 alloy-genesis.workspace = true
 alloy-hardforks.workspace = true
-alloy-network-primitives.workspace = true
 alloy-primitives.workspace = true
 alloy-rlp.workspace = true
 alloy-rpc-types-eth.workspace = true
 alloy-serde.workspace = true
+alloy-trie.workspace = true
 
 # op
 alloy-op-evm.workspace = true
 alloy-op-hardforks.workspace = true
+op-alloy-network.workspace = true
 op-alloy-rpc-types.workspace = true
 
 # revm
@@ -37,6 +38,7 @@ reth-trie-sparse = { workspace = true }
 
 # misc
 bincode.workspace = true
+eyre.workspace = true
 dyn-clone.workspace = true
 redb.workspace = true
 serde.workspace = true
